requirements.txt
kopf
kubernetes
loguru
PyYAML
jinja2
bcrypt
pykube
requests